assistente-juridico-multiagente/backend_api/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
crewai
crewai[tools]
google-generativeai
python-serper